requirements.txt
# Gap Assessment Agent - Complete Requirements
# Install with: pip install -r requirements.txt

# ============================================================================
# Google ADK and Gemini (REQUIRED)
# ============================================================================
google-adk>=0.1.0
google-generativeai>=0.3.0
langchain-google-genai>=1.0.0

# ============================================================================
# Vector Database (REQUIRED)
# ============================================================================
pinecone>=3.0.0
numpy>=1.24.0

# ============================================================================
# Embeddings (REQUIRED)
# ============================================================================
sentence-transformers>=2.2.0
torch>=2.0.0

# ============================================================================
# Text Processing (REQUIRED)
# ============================================================================
langchain>=0.1.0
langchain-text-splitters>=0.0.1
tiktoken>=0.5.0

# ============================================================================
# API Framework (REQUIRED for API endpoint)
# ============================================================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# ============================================================================
# Web Scraping (REQUIRED)
# ============================================================================
browser-use>=0.1.0
playwright>=1.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ============================================================================
# Document Parsing (REQUIRED)
# ============================================================================
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
openpyxl>=3.1.0

# ============================================================================
# Data Processing (REQUIRED)
# ============================================================================
pandas>=2.0.0

# ============================================================================
# Utilities (REQUIRED)
# ============================================================================
python-dateutil>=2.8.0

# ============================================================================
# Optional but Recommended
# ============================================================================
# For better performance with large datasets
# scipy>=1.11.0

# For additional logging capabilities
# structlog>=23.0.0

